PEP8 and project restructure
