+++ casca/bot.py
@@ -1,36 +1,38 @@
 # Discord Bot
+import logging
 import os
 import sys
+
 import discord
 import logbook
-import logging
+import yaml
+from discord.ext.commands import Bot
 from logbook import Logger
 from logbook import StreamHandler
 from logbook.compat import redirect_logging
-import yaml
-from discord.ext.commands import Bot
 
-extensions = ["Casca_Best_Bot.cogs.mathematics"]
+extensions = ["casca.cogs.mathematics"]
 
-Whitelisted_Servers = ["221708975698083841", #GERMAN LEARNING AND DISCUSSION
-                       "245333247796576257", #STEAMBOAT
-                       "293111771428945920", #BOT-TESTING
-                       "206935992022728704"] #HYPNOSIS
+Whitelisted_Servers = ["221708975698083841",  # GERMAN LEARNING AND DISCUSSION
+                       "245333247796576257",  # STEAMBOAT
+                       "293111771428945920",  # BOT-TESTING
+                       "206935992022728704"]  # HYPNOSIS
 
-Whitelisted_Channels = ["221708975698083841", #GERMAN LEARNING AND DISCUSSION: general
-                        "221709483284496394", #GERMAN LEARNING AND DISCUSSION: learning
-                        "222013061886640128", #GERMAN LEARNING AND DISCUSSION: deutsch-only
-                        "259006631185088516", #GERMAN LEARNING AND DISCUSSION: introductions
-                        "251115764680097794", #GERMAN LEARNING AND DISCUSSION: announcements
-                        "252121415912914946", #GERMAN LEARNING AND DISCUSSION: writing
-                        "248530603165614080", #GERMAN LEARNING AND DISCUSSION: botchannel
-                        "260865272292835329", #GERMAN LEARNING AND DISCUSSION: 0x1-bot
+Whitelisted_Channels = ["221708975698083841",  # GERMAN LEARNING AND DISCUSSION: general
+                        "221709483284496394",  # GERMAN LEARNING AND DISCUSSION: learning
+                        "222013061886640128",  # GERMAN LEARNING AND DISCUSSION: deutsch-only
+                        "259006631185088516",  # GERMAN LEARNING AND DISCUSSION: introductions
+                        "251115764680097794",  # GERMAN LEARNING AND DISCUSSION: announcements
+                        "252121415912914946",  # GERMAN LEARNING AND DISCUSSION: writing
+                        "248530603165614080",  # GERMAN LEARNING AND DISCUSSION: botchannel
+                        "260865272292835329",  # GERMAN LEARNING AND DISCUSSION: 0x1-bot
 
-                        "245333247796576257", #STEAMBOAT: general
+                        "245333247796576257",  # STEAMBOAT: general
 
-                        "293111771428945920", #BOT-TESTING: general
+                        "293111771428945920",  # BOT-TESTING: general
 
-                        "206935992022728704"] #HYPNOSIS: main
+                        "206935992022728704"]  # HYPNOSIS: main
+
 
 class Casca(Bot):
     def __init__(self, *args, **kwargs):
@@ -58,7 +60,8 @@ class Casca(Bot):
             return
 
         self.logger.info(
-            "LOADED Casca | LOGGED IN AS: {0.user.name}#{0.user.discriminator}.\n----------------------------------------------------------------------------------------------------".format(self))
+            "LOADED Casca | LOGGED IN AS: {0.user.name}#{0.user.discriminator}.\n----------------------------------------------------------------------------------------------------".format(
+                self))
 
         for cog in extensions:
             try:
@@ -81,15 +84,17 @@ class Casca(Bot):
         if message.channel.id not in Whitelisted_Channels:
             return
 
-        self.logger.info("MESSAGE: {message.content}".format(message=message, bot=" [BOT]" if message.author.bot else ""))
+        self.logger.info(
+            "MESSAGE: {message.content}".format(message=message, bot=" [BOT]" if message.author.bot else ""))
         self.logger.info("FROM: {message.author.name}".format(message=message))
 
         if message.server is not None:
             self.logger.info("CHANNEL: {message.channel.name}".format(message=message))
-            self.logger.info("SERVER: {0.server.name}\n----------------------------------------------------------------------------------------------------".format(message))
+            self.logger.info(
+                "SERVER: {0.server.name}\n----------------------------------------------------------------------------------------------------".format(
+                    message))
 
         await super().on_message(message)
-
 
     def run(self):
         try:
