+++ run.py
@@ -1,6 +1,6 @@
 from discord.ext.commands import when_mentioned_or
 
-from Casca_Best_Bot.bot import Casca
+from casca.bot import Casca
 
 if __name__ == "__main__":
     bot = Casca(command_prefix=when_mentioned_or("$"))
